# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clash Ren FF Tournament Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile (UID, in-game name)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/stats:
    get:
      operationId: getMyStats
      tags: [users]
      summary: Get current user stats summary (kills, wins, tournaments played)
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
          nullable: true
        googleId:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isProfileComplete:
          type: boolean
        inGameName:
          type: string
          nullable: true
        uid:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
        diamondBalance:
          type: integer
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - isProfileComplete
        - diamondBalance
        - isAdmin
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        inGameName:
          type: string
        uid:
          type: string
        profilePicture:
          type: string
          nullable: true
...
    UserStats:
      type: object
      properties:
        tournamentsPlayed:
          type: integer
        totalKills:
          type: integer
        totalWins:
          type: integer
        diamondsEarned:
          type: integer
        diamondsSpent:
          type: integer
        rank:
          type: integer
          nullable: true
      required:
        - tournamentsPlayed
        - totalKills
        - totalWins
        - diamondsEarned
        - diamondsSpent
...